backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cyclopts>=4.10.1",
    "fastmcp>=3.1.1",
    "mcp>=1.26.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.2",
]
